+++ Modules/SharePointDsc/Examples/Resources/SPSearchAuthoratativePage/1 - AuthoratativePage.ps1
@@ -1,0 +1,26 @@
+<#
+.EXAMPLE
+    This example shows how to create a Search Authoratative Page
+#>
+
+    Configuration Example 
+    {
+        param(
+            [Parameter(Mandatory = $true)]
+            [PSCredential]
+            $SetupAccount
+        )
+        Import-DscResource -ModuleName SharePointDsc
+
+        node localhost {
+            SPSearchAuthoratativePage AuthoratativePage
+            {
+                ServiceAppName       = "Search Service Application"
+                Path                 = "http://site.sharepoint.com/Pages/authoratative.aspx"
+                Action               = "Authoratative"
+                Level                = 0.0
+                Ensure               = "Present"
+                PsDscRunAsCredential = $SetupAccount
+            }
+        }
+    }